src/requirements.txt
opencv-python
numpy
pyzbar
openpyxl
pymodbus
Pillow
pysqlite3 